Pass 32-bit values into the printf.

+++ src/VBox/Devices/PC/BIOS/scsi.c
@@ -538,8 +538,8 @@ void scsi_enumerate_attached_devices(uint16_t io_base)
                     bios_dsk->devices[hd_index].lchs.cylinders = (uint16_t)cylinders;
 
                 BX_INFO("SCSI %d-ID#%d: LCHS=%u/%u/%u 0x%8X%8X sectors\n", devcount_scsi,
-                        i, (uint16_t)cylinders, heads, sectors_per_track, sectors >> 32,
-                        sectors & 0xffffffff);
+                        i, (uint16_t)cylinders, heads, sectors_per_track,
+                        (uint32_t)(sectors >> 32), (uint32_t)(sectors & 0xffffffff));
 
                 /* Write PCHS values. */
                 bios_dsk->devices[hd_index].pchs.heads = heads;
